reorganizing tiler utilities, deleting duplicates
