+++ solaris/tile/__init__.py
@@ -1,1 +1,1 @@
-from . import main, vector_utils
+from . import raster_tile, vector_tile
