+++ solaris/utils/tile.py
@@ -1,172 +1,13 @@
-from rasterio import Affine
-from rasterio import features
 import geopandas as gpd
 import numpy as np
 from shapely.geometry import box
 import rasterio
-from rasterio.vrt import WarpedVRT
-from rasterio.enums import Resampling
 from rasterio.io import DatasetReader
 from rasterio.warp import transform_bounds
+from rasterio.vrt import WarpedVRT
+from rasterio.enums import Resampling
 from rio_tiler.errors import RioTilerError
 from rasterio import transform
-import osr
-from solaris.utils.core import _check_gdf_load
-from solaris.utils.core import _check_rasterio_im_load
-
-
-def utm_getZone(longitude):
-    """Calculate UTM Zone from Longitude.
-
-    Arguments
-    ---------
-    longitude: float
-        longitude coordinate (Degrees.decimal degrees)
-
-    Returns
-    -------
-    out: int
-        UTM Zone number.
-
-    """
-
-    return (int(1+(longitude+180.0)/6.0))
-
-
-def utm_isNorthern(latitude):
-    """Determine if a latitude coordinate is in the northern hemisphere.
-
-    Arguments
-    ---------
-    latitude: float
-        latitude coordinate (Deg.decimal degrees)
-
-    Returns
-    -------
-    out: bool
-        ``True`` if `latitude` is in the northern hemisphere, ``False``
-        otherwise.
-
-    """
-
-    return (latitude > 0.0)
-
-
-def calculate_UTM_crs(coords):
-    """Calculate UTM Projection String.
-
-    Arguments
-    ---------
-    coords: list
-        ``[longitude, latitude]`` or
-        ``[min_longitude, min_latitude, max_longitude, max_latitude]`` .
-
-    Returns
-    -------
-    out: str
-        returns `proj4 projection string <https://proj4.org/usage/quickstart.html>`__
-
-    """
-
-    if len(coords) == 2:
-        longitude, latitude = coords
-    elif len(coords) == 4:
-        longitude = np.mean([coords[0], coords[2]])
-        latitude = np.mean([coords[1], coords[3]])
-
-    utm_zone = utm_getZone(longitude)
-
-    utm_isNorthern(latitude)
-
-    if utm_isNorthern(latitude):
-        direction_indicator = "+north"
-    else:
-        direction_indicator = "+south"
-
-    utm_crs = "+proj=utm +zone={} {} +ellps=WGS84 +datum=WGS84 +units=m +no_defs".format(utm_zone,
-                                                                                         direction_indicator)
-
-    return utm_crs
-
-
-def get_utm_vrt(source, crs='EPSG:3857', resampling=Resampling.bilinear,
-                src_nodata=None, dst_nodata=None):
-    """Get a :py:class:`rasterio.vrt.WarpedVRT` projection of a dataset.
-
-    Arguments
-    ---------
-    source : :py:class:`rasterio.io.DatasetReader`
-        The dataset to virtually warp using :py:class:`rasterio.vrt.WarpedVRT`.
-    crs : :py:class:`rasterio.crs.CRS`, optional
-        Coordinate reference system for the VRT. Defaults to 'EPSG:3857'
-        (Web Mercator).
-    resampling : :py:class:`rasterio.enums.Resampling` method, optional
-        Resampling method to use. Defaults to
-        :py:func:`rasterio.enums.Resampling.bilinear`. Alternatives include
-        :py:func:`rasterio.enums.Resampling.average`,
-        :py:func:`rasterio.enums.Resampling.cubic`, and others. See docs for
-        :py:class:`rasterio.enums.Resampling` for more information.
-    src_nodata : int or float, optional
-        Source nodata value which will be ignored for interpolation. Defaults
-        to ``None`` (all data used in interpolation).
-    dst_nodata : int or float, optional
-        Destination nodata value which will be ignored for interpolation.
-        Defaults to ``None``, in which case the value of `src_nodata` will be
-        used if provided, or ``0`` otherwise.
-
-    Returns
-    -------
-    A :py:class:`rasterio.vrt.WarpedVRT` instance with the transformation.
-
-    """
-
-    vrt_params = dict(
-        crs=crs,
-        resampling=Resampling.bilinear,
-        src_nodata=src_nodata,
-        dst_nodata=dst_nodata)
-
-    return WarpedVRT(source, **vrt_params)
-
-
-def get_utm_vrt_profile(source, crs='EPSG:3857',
-                        resampling=Resampling.bilinear,
-                        src_nodata=None, dst_nodata=None):
-    """Get a :py:class:`rasterio.profiles.Profile` for projection of a VRT.
-
-    Arguments
-    ---------
-    source : :py:class:`rasterio.io.DatasetReader`
-        The dataset to virtually warp using :py:class:`rasterio.vrt.WarpedVRT`.
-    crs : :py:class:`rasterio.crs.CRS`, optional
-        Coordinate reference system for the VRT. Defaults to ``"EPSG:3857"``
-        (Web Mercator).
-    resampling : :py:class:`rasterio.enums.Resampling` method, optional
-        Resampling method to use. Defaults to
-        ``rasterio.enums.Resampling.bilinear``. Alternatives include
-        ``rasterio.enums.Resampling.average``,
-        ``rasterio.enums.Resampling.cubic``, and others. See docs for
-        :py:class:`rasterio.enums.Resampling` for more information.
-    src_nodata : int or float, optional
-        Source nodata value which will be ignored for interpolation. Defaults
-        to ``None`` (all data used in interpolation).
-    dst_nodata : int or float, optional
-        Destination nodata value which will be ignored for interpolation.
-        Defaults to ``None``, in which case the value of `src_nodata`
-        will be used if provided, or ``0`` otherwise.
-
-    Returns
-    -------
-    A :py:class:`rasterio.profiles.Profile` instance with the transformation
-    applied.
-
-    """
-
-    with get_utm_vrt(source, crs=crs, resampling=resampling,
-                     src_nodata=src_nodata, dst_nodata=dst_nodata) as vrt:
-        vrt_profile = vrt.profile
-
-    return vrt_profile
 
 
 def tile_read_utm(source, bounds, tilesize, indexes=[1], nodata=None,
@@ -309,8 +150,8 @@ def get_wgs84_bounds(source):
         src = source
     else:
         src = rasterio.open(source)
-    wgs_bounds = transform_bounds(*[src.crs, 'epsg:4326'] +
-                                  list(src.bounds), densify_pts=21)
+    wgs_bounds = transform_bounds(*[src.crs, 'epsg:4326'] + list(src.bounds),
+                                  densify_pts=21)
     return wgs_bounds
 
 
@@ -340,52 +181,6 @@ def get_utm_bounds(source, utm_EPSG):
     utm_bounds = transform_bounds(*[src.crs, utm_EPSG] + list(src.bounds),
                                   densify_pts=21)
     return utm_bounds
-
-
-# Note, for mac osx compatability import something from shapely.geometry before
-# importing fiona or geopandas: https://github.com/Toblerity/Shapely/issues/553
-
-
-def read_vector_file(geoFileName):
-    """Read Fiona-Supported Files into GeoPandas GeoDataFrame.
-
-    Warning
-    ----
-    This will raise an exception for empty GeoJSON files, which GDAL and Fiona
-    cannot read. ``try/except`` the :py:exc:`Fiona.errors.DriverError` or
-    :py:exc:`Fiona._err.CPLE_OpenFailedError` if you must use this.
-
-    """
-
-    return gpd.read_file(geoFileName)
-
-
-def transformToUTM(gdf, utm_crs, estimate=True, calculate_sindex=True):
-    """Transform GeoDataFrame to UTM coordinate reference system.
-
-    Arguments
-    ---------
-    gdf : :py:class:`geopandas.GeoDataFrame`
-        :py:class:`geopandas.GeoDataFrame` to transform.
-    utm_crs : str
-        :py:class:`rasterio.crs.CRS` string for destination UTM CRS.
-    estimate : bool, optional
-        .. deprecated:: 0.2.0
-            This argument is no longer used.
-    calculate_sindex : bool, optional
-        .. deprecated:: 0.2.0
-            This argument is no longer used.
-
-    Returns
-    -------
-    gdf : :py:class:`geopandas.GeoDataFrame`
-        The input :py:class:`geopandas.GeoDataFrame` converted to
-        `utm_crs` coordinate reference system.
-
-    """
-
-    gdf = gdf.to_crs(utm_crs)
-    return gdf
 
 
 def search_gdf_bounds(gdf, tile_bounds):
@@ -440,47 +235,6 @@ def search_gdf_polygon(gdf, tile_polygon):
     if precise_matches.empty:
         precise_matches = gpd.GeoDataFrame(geometry=[])
     return precise_matches
-
-
-def vector_tile_utm(gdf, tile_bounds, min_partial_perc=0.1,
-                    geom_type="Polygon", use_sindex=True):
-    """Wrapper for :func:`clip_gdf` that converts `tile_bounds` to a polygon.
-
-    Arguments
-    ---------
-    gdf : :class:`geopandas.GeoDataFrame`
-        A :py:class:`geopandas.GeoDataFrame` of polygons to clip.
-    tile_bounds : list-like of floats
-        :obj:`list` of shape ``(W, S, E, N)`` denoting the boundaries of an
-        imagery tile. Converted to a polygon for :func:`clip_gdf`.
-    min_partial_perc : float
-        The minimum fraction of an object in `gdf` that must be
-        preserved. Defaults to 0.0 (include any object if any part remains
-        following clipping).
-    use_sindex : bool, optional
-        Use the `gdf` sindex be used for searching. Improves efficiency
-        but requires `libspatialindex <http://libspatialindex.github.io/>`__ .
-
-    Returns
-    -------
-    small_gdf : :py:class:`geopandas.GeoDataFrame`
-        `gdf` with all contained objects clipped to `tile_bounds`.
-        See notes above for details on additional clipping columns added.
-    """
-    tile_polygon = box(*tile_bounds)
-    small_gdf = clip_gdf(gdf, tile_polygon,
-                         min_partial_perc=min_partial_perc,
-                         geom_type=geom_type
-                         )
-
-    return small_gdf
-
-
-def getCenterOfGeoFile(gdf, estimate=True):
-
-    # TODO implement calculate UTM from gdf  see osmnx implementation
-
-    pass
 
 
 def clip_gdf(gdf, poly_to_cut, min_partial_perc=0.0, geom_type="Polygon",
@@ -567,88 +321,3 @@ def clip_gdf(gdf, poly_to_cut, min_partial_perc=0.0, geom_type="Polygon",
     # TODO: IMPLEMENT TRUNCATION MEASUREMENT FOR LINESTRINGS
 
     return cutGeoDF
-
-
-def rasterize_gdf(gdf, src_shape, burn_value=1,
-                  src_transform=Affine(1.0, 0.0, 0.0, 0.0, 1.0, 0.0)):
-    """Convert a GeoDataFrame to a binary image (array) mask.
-
-    Uses :py:func:`rasterio.features.rasterize` to generate a raster mask from
-    object geometries in `gdf` .
-
-    Arguments
-    ---------
-    gdf : :py:class:`geopandas.GeoDataFrame`
-        A :py:class:`geopandas.GeoDataFrame` of objects to convert into a mask.
-    src_shape : list-like of 2 ints
-        Shape of the output array in ``(Y, X)`` pixel units.
-    burn_value : int in range(0, 255), optional
-        Integer value for pixels corresponding to objects from `gdf` .
-        Defaults to 1.
-    src_transform : :py:class:`affine.Affine`, optional
-        Affine transformation for the output raster. If not provided, defaults
-        to arbitrary pixel units.
-
-    Returns
-    -------
-    img : :class:`np.ndarray`, dtype ``uint8``
-        A NumPy array of integers with 0s where no pixels from objects in
-        `gdf` exist, and `burn_value` where they do. Shape is
-        defined by `src_shape`.
-    """
-    if not gdf.empty:
-        img = features.rasterize(
-            ((geom, burn_value) for geom in gdf.geometry),
-            out_shape=src_shape,
-            transform=src_transform
-        )
-    else:
-        img = np.zeros(src_shape).astype(np.uint8)
-
-    return img
-
-
-def vector_gdf_get_projection_unit(vector_file):
-    """Get the projection unit for a vector_file or gdf.
-
-    Arguments
-    ---------
-    vector_file : :py:class:`geopandas.GeoDataFrame` or geojson/shapefile
-        A vector file or gdf with georeferencing
-
-    Returns
-    -------
-    unit : String
-        The unit i.e. meters or degrees, of the projection
-    """
-    c = _check_gdf_load(vector_file)
-    crs = c.crs
-    srs = osr.SpatialReference()
-    x = (crs['init']).split(":")[1]
-    srs.ImportFromEPSG(int(x))
-    WKT = srs.ExportToWkt()
-    unit = WKT.split("UNIT[")[1].split(",")[0]
-    return unit
-
-
-def raster_get_projection_unit(image):
-    """Get the projection unit for a vector_file.
-
-    Arguments
-    ---------
-    image : raster image, GeoTIFF or other format
-        A raster file with georeferencing
-
-    Returns
-    -------
-    unit : String
-        The unit i.e. meters or degrees, of the projection
-    """
-    c = _check_rasterio_im_load(image)
-    crs = c.crs
-    srs = osr.SpatialReference()
-    x = (crs['init']).split(":")[1]
-    srs.ImportFromEPSG(int(x))
-    WKT = srs.ExportToWkt()
-    unit = WKT.split("UNIT[")[1].split(",")[0]
-    return unit
